+++ src/main/java/dev/nilswitt/webmap/api/MapItemController.java
@@ -2,11 +2,11 @@ package dev.nilswitt.webmap.api;
 
 import dev.nilswitt.webmap.api.exceptions.ForbiddenException;
 import dev.nilswitt.webmap.api.exceptions.MapItemNotFoundException;
-import dev.nilswitt.webmap.api.helpers.ApiAuthorizationHelper;
 import dev.nilswitt.webmap.entities.MapItem;
 import dev.nilswitt.webmap.entities.SecurityGroup;
 import dev.nilswitt.webmap.entities.User;
 import dev.nilswitt.webmap.entities.repositories.MapItemRepository;
+import dev.nilswitt.webmap.security.PermissionUtil;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -25,9 +25,11 @@ public class MapItemController {
 
     private final MapItemRepository repository;
     private final MapItemModelAssembler assembler;
+    private  final PermissionUtil permissionUtil;
 
-    public MapItemController(MapItemRepository repository, MapItemModelAssembler assembler) {
+    public MapItemController(MapItemRepository repository, MapItemModelAssembler assembler, PermissionUtil permissionUtil) {
         this.repository = repository;
+        this.permissionUtil = permissionUtil;
         this.assembler = assembler;
     }
 
@@ -35,27 +37,24 @@ public class MapItemController {
     // tag::get-aggregate-root[]
     @GetMapping("")
     CollectionModel<EntityModel<MapItem>> all(@AuthenticationPrincipal User userDetails) {
-        try {
-            ApiAuthorizationHelper.requireAnyScope(userDetails, SecurityGroup.UserRoleTypeEnum.MAPITEM,
-                    "User does not have permission to view map items.", SecurityGroup.UserRoleScopeEnum.VIEW);
+        if (this.permissionUtil.hasAccess(userDetails, SecurityGroup.UserRoleScopeEnum.VIEW, SecurityGroup.UserRoleTypeEnum.MAPITEM)) {
 
-            List<EntityModel<MapItem>> items = this.repository.findAll().stream()
+            List<EntityModel<MapItem>> entities = this.repository.findAll().stream()
                     .map(this.assembler::toModel)
                     .collect(Collectors.toList());
-
-            return CollectionModel.of(items, linkTo(methodOn(MapItemController.class).all(null)).withSelfRel());
-        } catch (ForbiddenException e) {
-            return CollectionModel.of(List.of(), linkTo(methodOn(MapItemController.class).all(null)).withSelfRel());
+            return CollectionModel.of(entities, linkTo(methodOn(MapItemController.class).all(null)).withSelfRel());
         }
+
+        return CollectionModel.of(this.permissionUtil.getMapItemsForUser(userDetails).stream().map(this.assembler::toModel).collect(Collectors.toList()), linkTo(methodOn(MapItemController.class).all(null)).withSelfRel());
 
     }
     // end::get-aggregate-root[]
 
     @PostMapping("")
     EntityModel<MapItem> newEntity(@RequestBody MapItem newEntity, @AuthenticationPrincipal User userDetails) {
-        ApiAuthorizationHelper.requireAnyScope(userDetails, SecurityGroup.UserRoleTypeEnum.MAPITEM,
-                "User does not have permission to create map items.", SecurityGroup.UserRoleScopeEnum.CREATE);
-
+        if (!this.permissionUtil.hasAccess(userDetails, SecurityGroup.UserRoleScopeEnum.CREATE, SecurityGroup.UserRoleTypeEnum.MAPITEM)) {
+            throw new ForbiddenException("User does not have permission to create overlays.");
+        }
         return this.assembler.toModel(this.repository.save(newEntity));
     }
 
@@ -63,9 +62,10 @@ public class MapItemController {
 
     @GetMapping("{id}")
     EntityModel<MapItem> one(@PathVariable UUID id, @AuthenticationPrincipal User userDetails) {
-        ApiAuthorizationHelper.requireAnyScope(userDetails, SecurityGroup.UserRoleTypeEnum.MAPITEM,
-                "User does not have permission to view map items.", SecurityGroup.UserRoleScopeEnum.VIEW);
-
+        MapItem entity = this.repository.findById(id).orElseThrow(() -> new MapItemNotFoundException(id));
+        if (this.permissionUtil.hasAccess(userDetails, SecurityGroup.UserRoleScopeEnum.VIEW, entity)) {
+            throw new ForbiddenException("User does not have permission to view overlays.");
+        }
         return this.assembler.toModel(
                 this.repository.findById(id)
                         .orElseThrow(() -> new MapItemNotFoundException(id))
@@ -74,11 +74,11 @@ public class MapItemController {
 
     @PutMapping("{id}")
     EntityModel<MapItem> replaceEntity(@RequestBody MapItem newEntity, @PathVariable UUID id, @AuthenticationPrincipal User userDetails) {
-        ApiAuthorizationHelper.requireAnyScope(userDetails, SecurityGroup.UserRoleTypeEnum.MAPITEM,
-                "User does not have permission to edit map items.", SecurityGroup.UserRoleScopeEnum.EDIT);
+        MapItem entity = this.repository.findById(id).orElseThrow(() -> new MapItemNotFoundException(id));
 
-        MapItem entity = this.repository.findById(id)
-                .orElseThrow(() -> new MapItemNotFoundException(id));
+        if (!this.permissionUtil.hasAccess(userDetails, SecurityGroup.UserRoleScopeEnum.EDIT, entity)) {
+            throw new ForbiddenException("User does not have permission to edit overlays.");
+        }
 
         entity.setName(newEntity.getName());
         entity.setPosition(newEntity.getPosition());
@@ -88,9 +88,11 @@ public class MapItemController {
 
     @DeleteMapping("{id}")
     void deleteEntity(@PathVariable UUID id, @AuthenticationPrincipal User userDetails) {
-        ApiAuthorizationHelper.requireAnyScope(userDetails, SecurityGroup.UserRoleTypeEnum.MAPITEM,
-                "User does not have permission to delete map items.", SecurityGroup.UserRoleScopeEnum.DELETE);
+        MapItem entity = this.repository.findById(id).orElseThrow(() -> new MapItemNotFoundException(id));
 
+        if (!this.permissionUtil.hasAccess(userDetails, SecurityGroup.UserRoleScopeEnum.DELETE, entity)) {
+            throw new ForbiddenException("User does not have permission to delete overlays.");
+        }
         this.repository.deleteById(id);
     }
 }
