+++ src/main/java/dev/nilswitt/webmap/api/UnitController.java
@@ -2,11 +2,11 @@ package dev.nilswitt.webmap.api;
 
 import dev.nilswitt.webmap.api.exceptions.ForbiddenException;
 import dev.nilswitt.webmap.api.exceptions.UnitNotFoundException;
-import dev.nilswitt.webmap.api.helpers.ApiAuthorizationHelper;
 import dev.nilswitt.webmap.entities.SecurityGroup;
 import dev.nilswitt.webmap.entities.Unit;
 import dev.nilswitt.webmap.entities.User;
 import dev.nilswitt.webmap.entities.repositories.UnitRepository;
+import dev.nilswitt.webmap.security.PermissionUtil;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.hateoas.CollectionModel;
@@ -27,37 +27,33 @@ public class UnitController {
 
     private final UnitRepository repository;
     private final UnitModelAssembler assembler;
+    private final PermissionUtil permissionUtil;
     private final Logger logger = LogManager.getLogger(this.getClass());
 
-    public UnitController(UnitRepository userRepository, UnitModelAssembler assembler) {
+    public UnitController(UnitRepository userRepository, UnitModelAssembler assembler, PermissionUtil permissionUtil) {
         this.repository = userRepository;
         this.assembler = assembler;
+        this.permissionUtil = permissionUtil;
     }
 
-    // Aggregate root
-    // tag::get-aggregate-root[]
     @GetMapping("")
     CollectionModel<EntityModel<Unit>> all(@AuthenticationPrincipal User userDetails) {
+        if (this.permissionUtil.hasAccess(userDetails, SecurityGroup.UserRoleScopeEnum.VIEW, SecurityGroup.UserRoleTypeEnum.UNIT)) {
 
-        try {
-            ApiAuthorizationHelper.requireAnyScope(userDetails, SecurityGroup.UserRoleTypeEnum.UNIT,
-                    "User does not have permission to create users.", SecurityGroup.UserRoleScopeEnum.VIEW);
-            List<EntityModel<Unit>> units = this.repository.findAll().stream()
+            List<EntityModel<Unit>> entities = this.repository.findAll().stream()
                     .map(this.assembler::toModel)
                     .collect(Collectors.toList());
-            return CollectionModel.of(units, linkTo(methodOn(UnitController.class).all(null)).withSelfRel());
-
-        } catch (ForbiddenException e) {
-            return CollectionModel.of(List.of(), linkTo(methodOn(UnitController.class).all(null)).withSelfRel());
-
+            return CollectionModel.of(entities, linkTo(methodOn(UnitController.class).all(null)).withSelfRel());
         }
+
+        return CollectionModel.of(this.permissionUtil.getUnitsForUser(userDetails).stream().map(this.assembler::toModel).collect(Collectors.toList()), linkTo(methodOn(UnitController.class).all(null)).withSelfRel());
     }
-    // end::get-aggregate-root[]
 
     @PostMapping("")
     EntityModel<Unit> newEntity(@RequestBody Unit newEntity, @AuthenticationPrincipal User userDetails) {
-        ApiAuthorizationHelper.requireAnyScope(userDetails, SecurityGroup.UserRoleTypeEnum.UNIT,
-                "User does not have permission to create users.", SecurityGroup.UserRoleScopeEnum.CREATE);
+        if (!this.permissionUtil.hasAccess(userDetails, SecurityGroup.UserRoleScopeEnum.CREATE, SecurityGroup.UserRoleTypeEnum.UNIT)) {
+            throw new ForbiddenException("User does not have permission to create overlays.");
+        }
         return this.assembler.toModel(this.repository.save(newEntity));
     }
 
@@ -65,8 +61,10 @@ public class UnitController {
 
     @GetMapping("{id}")
     EntityModel<Unit> one(@PathVariable UUID id, @AuthenticationPrincipal User userDetails) {
-        ApiAuthorizationHelper.requireAnyScope(userDetails, SecurityGroup.UserRoleTypeEnum.UNIT,
-                "User does not have permission to view units.", SecurityGroup.UserRoleScopeEnum.VIEW);
+        Unit entity = this.repository.findById(id).orElseThrow(() -> new UnitNotFoundException(id));
+        if (this.permissionUtil.hasAccess(userDetails, SecurityGroup.UserRoleScopeEnum.VIEW, entity)) {
+            throw new ForbiddenException("User does not have permission to view overlays.");
+        }
         return this.assembler.toModel(
                 this.repository.findById(id)
                         .orElseThrow(() -> new UnitNotFoundException(id))
@@ -75,10 +73,11 @@ public class UnitController {
 
     @PutMapping("{id}")
     EntityModel<Unit> replaceEntity(@RequestBody Unit newEntity, @PathVariable UUID id, @AuthenticationPrincipal User userDetails) {
-        ApiAuthorizationHelper.requireAnyScope(userDetails, SecurityGroup.UserRoleTypeEnum.UNIT,
-                "User does not have permission to edit unit.", SecurityGroup.UserRoleScopeEnum.EDIT);
-        Unit entity = this.repository.findById(id)
-                .orElseThrow(() -> new UnitNotFoundException(id));
+        Unit entity = this.repository.findById(id).orElseThrow(() -> new UnitNotFoundException(id));
+
+        if (!this.permissionUtil.hasAccess(userDetails, SecurityGroup.UserRoleScopeEnum.EDIT, entity)) {
+            throw new ForbiddenException("User does not have permission to edit overlays.");
+        }
 
         entity.setName(newEntity.getName());
         entity.setPosition(newEntity.getPosition());
@@ -90,8 +89,11 @@ public class UnitController {
 
     @DeleteMapping("{id}")
     void deleteEntity(@PathVariable UUID id, @AuthenticationPrincipal User userDetails) {
-        ApiAuthorizationHelper.requireAnyScope(userDetails, SecurityGroup.UserRoleTypeEnum.UNIT,
-                "User does not have permission to delete units.", SecurityGroup.UserRoleScopeEnum.DELETE);
+        Unit entity = this.repository.findById(id).orElseThrow(() -> new UnitNotFoundException(id));
+
+        if (!this.permissionUtil.hasAccess(userDetails, SecurityGroup.UserRoleScopeEnum.DELETE, entity)) {
+            throw new ForbiddenException("User does not have permission to delete overlays.");
+        }
         this.repository.deleteById(id);
     }
 }
